🐞 fix: renaming the routs and deps paths
