+++ app/components/credetialTracker/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Box, Typography, Paper, styled, Card, CardContent } from '@mui/material'
-import { FormData } from '../../credentialForm/form/types/Types'
+import { FormData } from '../../skillform/form/types/Types'
 import { Logo } from '../../Assets/SVGs'
 
 // Styled components
